+++ src/key/generics.rs
@@ -1,6 +1,43 @@
 use hkdf::Hkdf;
 use sha2::Sha256;
-use super::TryPublicKeyFrom;
+
+/// Attempt a value -> `PublicKey` conversion that consumes the input value. The opposite of [TryIntoPublicKey].
+pub trait TryPublicKeyFrom<T>: Key + Sized {
+    fn try_pk_from(x: T) -> Result<Self, ()>;
+}
+
+/// Attempt a value -> `PublicKey` conversion that consumes the input value. The opposite of [TryPublicKeyFrom].
+pub trait TryIntoPublicKey<U: Key> {
+    fn try_into_pk(self) -> Result<U, ()>;
+}
+
+impl<T, U> TryIntoPublicKey<U> for T
+    where
+        U: TryPublicKeyFrom<T>
+{
+    fn try_into_pk(self) -> Result<U, ()> {
+        U::try_pk_from(self)
+    }
+}
+
+/// Attempt a value -> `SecretKey` conversion that consumes the input value. The opposite of [TryIntoSecretKey].
+pub trait TrySecretKeyFrom<T>: Key {
+    fn try_sk_from(x: T) -> Result<Self::SecretKey, ()>;
+}
+
+/// Attempt a value -> `SecretKey` conversion that consumes the input value. The opposite of [TrySecretKeyFrom].
+pub trait TryIntoSecretKey<U: Key> {
+    fn try_into_sk(self) -> Result<U::SecretKey, ()>;
+}
+
+impl<T, U> TryIntoSecretKey<U> for T
+    where
+        U: TrySecretKeyFrom<T>
+{
+    fn try_into_sk(self) -> Result<U::SecretKey, ()> {
+        U::try_sk_from(self)
+    }
+}
 
 pub trait Key {
     const EC_KEY_LEN: usize;
